Move everything from block_printer.rs to cfg_printer.rs
